Remove 128-bit integer variants from Number
