+++ src/value/de.rs
@@ -36,29 +36,11 @@ impl<'de> Deserialize<'de> for Value {
                 Ok(Value::Number(i.into()))
             }
 
-            serde_if_integer128! {
-                fn visit_i128<E>(self, i: i128) -> Result<Value, E>
-                where
-                    E: SError,
-                {
-                    Ok(Value::Number(i.into()))
-                }
-            }
-
             fn visit_u64<E>(self, u: u64) -> Result<Value, E>
             where
                 E: SError,
             {
                 Ok(Value::Number(u.into()))
-            }
-
-            serde_if_integer128! {
-                fn visit_u128<E>(self, u: u128) -> Result<Value, E>
-                where
-                    E: SError,
-                {
-                    Ok(Value::Number(u.into()))
-                }
             }
 
             fn visit_f64<E>(self, f: f64) -> Result<Value, E>
