+++ tests/test_serde.rs
@@ -14,7 +14,9 @@
 #[macro_use]
 extern crate serde_derive;
 
+#[macro_use]
 extern crate serde;
+
 extern crate serde_yaml;
 
 extern crate unindent;
@@ -91,52 +93,28 @@ fn test_int_max_i64() {
     test_serde(&thing, &yaml);
 }
 
-#[cfg(integer128)]
-#[test]
-fn test_i128_small() {
-    let thing: i128 = -256;
-    let yaml = unindent(
-        "
-        ---
-        -256",
-    );
-    test_serde(&thing, &yaml);
-}
+serde_if_integer128! {
+    #[test]
+    fn test_i128_small() {
+        let thing: i128 = -256;
+        let yaml = unindent(
+            "
+            ---
+            -256",
+        );
+        test_serde(&thing, &yaml);
+    }
 
-#[cfg(integer128)]
-#[test]
-fn test_i128_big() {
-    let thing: i128 = ::std::i64::MIN as i128 - 1;
-    let yaml = unindent(
-        "
-        ---
-        -9223372036854775809",
-    );
-    test_serde(&thing, &yaml);
-}
-
-#[cfg(integer128)]
-#[test]
-fn test_u128_small() {
-    let thing: u128 = 256;
-    let yaml = unindent(
-        "
-        ---
-        256",
-    );
-    test_serde(&thing, &yaml);
-}
-
-#[cfg(integer128)]
-#[test]
-fn test_u128_big() {
-    let thing: u128 = ::std::u64::MAX as u128 + 1;
-    let yaml = unindent(
-        "
-        ---
-        18446744073709551616",
-    );
-    test_serde(&thing, &yaml);
+    #[test]
+    fn test_u128_small() {
+        let thing: u128 = 256;
+        let yaml = unindent(
+            "
+            ---
+            256",
+        );
+        test_serde(&thing, &yaml);
+    }
 }
 
 #[test]
